verify JWT

+++ api/controller/post.js
@@ -10,10 +10,13 @@ export const getPosts = (req, res) => {
 };
 
 export const getPost = (req, res) => {
-  const q = "SELECT * FROM posts WHERE id = ?";
+  const q =
+    "SELECT p.id, `username`, `title`, `desc`, p.img, u.img AS userImg, `cat`,`date` FROM users u JOIN posts p ON u.id = p.uid WHERE p.id = ? ";
+
   db.query(q, [req.params.id], (err, data) => {
     if (err) return res.status(500).json(err);
-    return res.status(200).json(data);
+
+    return res.status(200).json(data[0]);
   });
 };
 
@@ -34,10 +37,20 @@ export const addPost = (req, res) => {
 };
 
 export const deletePost = (req, res) => {
-  const q = "DELETE FROM posts WHERE id = ?";
-  db.query(q, [req.params.id], (err, data) => {
-    if (err) return res.status(500).json(err);
-    return res.status(200).json("Post deleted successfully");
+  const token = req.cookies.access_token;
+  if (!token) return res.status(401).json("Not authenticated!");
+
+  jwt.verify(token, "jwtkey", (err, userInfo) => {
+    if (err) return res.status(403).json("Token is not valid!");
+
+    const postId = req.params.id;
+    const q = "DELETE FROM posts WHERE `id` = ? AND `uid` = ?";
+
+    db.query(q, [postId, userInfo.id], (err, data) => {
+      if (err) return res.status(403).json("You can delete only your post!");
+
+      return res.json("Post has been deleted!");
+    });
   });
 };
